Sort more things alphabetically
